Move zod validation schemas
